+++ WalletSystem/InputPayment.cs
@@ -5,6 +5,6 @@ namespace WalletSystem
     public class InputPayment:Payment
     {
         public InputPurpose purpose { get; private set; }
-        public InputPayment( double tempSum, InputPurpose pur) : base( tempSum) { purpose = pur; }
+        public InputPayment( decimal tempSum, InputPurpose pur) : base( tempSum) { purpose = pur; }
     }
 }
